fix(test): use custom testEnvironment to patch ReadableStream before jest init
setupFiles runs AFTER jest-environment-node's constructor — too late.
The ReadableStream \"readonly property\" crash happens inside the constructor
itself, before any user code runs.

Replace jest.setup.ts (setupFiles) with src/jest.env.js (custom
testEnvironment) that patches non-configurable/non-writable Web Streams
globals at module-load time, before NodeEnvironment is instantiated.
Covers both value-property and getter-only variants to handle Node 18–22.

Fixes #43
